Refactor connectivity tests for integration tests
